refactor(website): refactor to single component and make pink (#14074)
