+++ packages/website/pages/docs/reference/rpc-configuration.mdx
@@ -1,10 +1,10 @@
-import ConnectToMetamask from 'components/MetamaskNetworkButton';
+import ConnectToMetamaskButton from 'components/ConnectToMetamaskButton';
 
 # RPC configuration
 
 ## Sepolia
-
-<ConnectToMetamask buttonText={"Connect to Sepolia"} />
+<br/>
+<ConnectToMetamaskButton network="Sepolia" />
 
 | Name               | Value                        |
 | ------------------ | ---------------------------- |
@@ -14,8 +14,8 @@ import ConnectToMetamask from 'components/MetamaskNetworkButton';
 | Block Explorer URL | https://sepolia.etherscan.io |
 
 ## Taiko
-
-<ConnectToMetamask buttonText={"Connect to Taiko"} />
+<br/>
+<ConnectToMetamaskButton network="Taiko" />
 
 | Name               | Value                           |
 | ------------------ | ------------------------------- |
